applied text printing animation to cinema; game manager controls state of game, linked cinema, main menu, and game
cinema now plays scene with respective narrative script printing char by char. if it is the end of a sentence, there is a longer delay. created a game manager that controls the state of the game and disables/enables the relevant components. pressing play from main menu now goes into gameplay scene