normal ghost ai freaking works, hell yeah
entering the room will make the ghost appear only if it is nighttime. there is a brief interval when ghost is not active once player enters room so that it doesn't accidentally kill player immediately if ghost is spawned near doorway. ghost wanders when player is outside detection radius and follows when within detection. returns back to spawn state once player leaves room